requirements.txt
asgiref==3.2.10
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
Django==3.1.2
google-api-core==1.22.4
google-api-python-client==1.12.3
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-cloud-core==1.4.3
google-cloud-storage==1.31.2
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
mysql-connector-python==8.0.21
oauthlib==3.1.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.10
